Add new projection
