+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod types;
 pub mod models;
 pub mod utils;
+pub mod traits;
 pub mod projections;
-pub mod geometries;
+pub mod shape;
