+++ src/projections/mod.rs
@@ -1,3 +1,3 @@
 mod constants;
 mod pscp;
-pub mod vgcp;
+mod vgcp;
