+++ src/projections/vgcp/mod.rs
@@ -1,0 +1,2 @@
+mod ico_vgcp;
+mod vgcp;
