+++ src/traits/mod.rs
@@ -1,0 +1,2 @@
+pub mod projection;
+pub mod polyhedron;
