+++ src/traits/projection.rs
@@ -1,0 +1,8 @@
+use crate::models::common::Position;
+
+use super::polyhedron::Polyhedron;
+
+pub trait Projection {
+    fn forward(positions: Vec<Position>, shape: &dyn Polyhedron) -> [f64; 2];
+    fn inverse(&self) -> String;
+}
